Fix maven and iOS setup
