+++ .github/workflows/kotlin-ci.yml
@@ -46,3 +46,7 @@ jobs:
 
     - name: Build Kotlin Multiplatform
       run: tools/kotlin-build
+
+    - name: Build KMP Sample
+      run: ./gradlew assemble
+      working-directory: samples/kmp
