+++ samples/kmp/iosApp/iosApp/ContentView.swift
@@ -2,7 +2,7 @@ import SwiftUI
 import shared
 
 struct ContentView: View {
-	let greet = Greeting().greet()
+	let greet = Sample().greet()
 
 	var body: some View {
 		Text(greet)
