+++ samples/kmp/shared/src/androidMain/kotlin/com/example/kmpsample/Platform.kt
@@ -1,6 +1,9 @@
 package com.example.kmpsample
 
 class AndroidPlatform : Platform {
+    init {
+        System.loadLibrary("TrustWalletCore")
+    }
     override val name: String = "Android ${android.os.Build.VERSION.SDK_INT}"
 }
