--- samples/kmp/shared/src/commonMain/kotlin/com/example/kmpsample/Greeting.kt
@@ -1,9 +1,0 @@
-package com.example.kmpsample
-
-class Greeting {
-    private val platform: Platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
